+++ src/lib/output/themes/default/partials/typeDetails.tsx
@@ -309,6 +309,7 @@ function renderIndexSignature(context: DefaultThemeRenderContext, index: Signatu
     return (
         <li class="tsd-parameter-index-signature">
             <h5>
+                {index.flags.isReadonly && <span class="tsd-signature-keyword">readonly </span>}
                 <span class="tsd-signature-symbol">[</span>
                 {index.parameters!.map((item) => (
                     <>