+++ src/lib/output/themes/default/templates/reflection.tsx
@@ -84,6 +84,7 @@ function renderIndexSignature(context: DefaultThemeRenderContext, index: Signatu
     return (
         <li class="tsd-index-signature">
             <div class="tsd-signature">
+                {index.flags.isReadonly && <span class="tsd-signature-keyword">readonly </span>}
                 <span class="tsd-signature-symbol">[</span>
                 {index.parameters!.map((item) => (
                     <>