+++ src/test/converter/types/index-signature.ts
@@ -8,4 +8,21 @@ export interface PartialIndex {
 
 export interface UnionIndex {
     [optName: string | symbol]: unknown;
+    [numName: number]: string;
+}
+
+export interface ReadonlyIndex {
+    readonly [x: string]: string;
+}
+
+declare const symbolMethodName: symbol;
+declare const symbolPropertyName: symbol;
+
+export class A {
+    [symbolMethodName]() {
+        return 1;
+    }
+    [symbolPropertyName]() {
+        return "x";
+    }
 }